Modify SLA state based on approvals
